requirements.txt
flask
Flask-SQLAlchemy
marshmallow-sqlalchemy
flask_marshmallow
psycopg2-binary
python-dotenv
pyjwt